+++ src/screens/Details/index.tsx
